# api/openapi.yaml
openapi: "3.1.0"
info:
  title: qimeng-media API
  version: "0.1.0"
  description: |
    绮梦影库 NAS 多端版服务端协议（协议宪法）。
    本文件是三端（Go 服务端 / TS Web / Kotlin Android）SDK 的唯一生成源。
    规则：
    1. 先改本文件 → make sdk 重新生成 → 按编译错误适配代码。禁止手写客户端。
    2. 所有路径带 /api/v1 版本前缀；破坏性变更必须升版本或加兼容字段。
    3. 分页统一 cursor 式；错误统一 Error 结构。
    4. 领域语义（推荐/统计/匹配）见 docs/DOMAIN_RULES.md，本文件只定义传输结构。
servers:
  - url: "{scheme}://{host}:{port}"
    variables:
      scheme: { default: http, enum: [http, https] }
      host:   { default: nas.local }
      port:   { default: "8420" }
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/setup:
    post:
      summary: 首次初始化（设置管理密码，返回一次性显示的 token）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthSetupRequest" }
      responses:
        "201": { description: 创建成功，返回 token（仅此一次明文显示）, content: { application/json: { schema: { $ref: "#/components/schemas/AuthToken" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /media/orig/{assetId}:
    get:
      summary: 原图/原视频直链（HTTP Range 支持；签名参数校验；永不转码）
      security: []
      parameters:
        - { $ref: "#/components/parameters/AssetId" }
        - { name: exp, in: query, required: true, schema: { type: integer }, description: 过期时间戳 }
        - { name: sig, in: query, required: true, schema: { type: string }, description: HMAC 签名 }
      responses:
        "200": { description: 文件内容（Content-Type 按实际类型） }
        "206": { description: Range 分片 }
        "403": { description: 签名无效或过期 }
  /media/thumb/{assetId}:
    get:
      summary: 缩略图（生成策略见 DOMAIN_RULES §11；缓存 immutable，缓存键含 size）
      security: []
      parameters:
        - { $ref: "#/components/parameters/AssetId" }
        - { name: size, in: query, schema: { type: string, enum: [sm, md, lg], default: md }, description: 网格/列表/大图三档（缓存键组成部分）；像素档位单一来源 server/internal/thumbnail/cachekey.go 的 SizeSmall/SizeGrid/SizePreview 常量（md 档像素由 Thumbnail 配置 LongSide 决定，未配置回落 SizeGrid），改动双同步 }
        - { name: exp, in: query, required: true, schema: { type: integer } }
        - { name: sig, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: WebP 缩略图 }

  # ---------- 上传与文件管理 ----------
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: "QM token" }
  parameters:
...
    AssetId:   { name: assetId,   in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    Conflict:    { description: 冲突,     content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, description: 机器可读错误码（PATH_ESCAPE/TOO_LARGE/INVALID_TYPE/...） }
        message: { type: string, description: 人类可读信息（不含内部路径） }
    AuthSetupRequest:
      type: object
      required: [password]
      properties:
        password: { type: string, minLength: 8 }
    AuthToken:
      type: object
      properties:
        token: { type: string, description: 仅 setup 响应中明文返回一次 }
...